Refactor: 메소드명 수정
